Revert "Issue #223: separates things common to cs-ledger and cs-blockchain into a new package cs-core"
This reverts commit 893cf15604b9663eda243cb604b4d4f4e67626cb.
