+++ pkgs.nix
@@ -15,7 +15,6 @@ import nixpkgs {
                 enableParallelBuilding = false;
               });
               cs-blockchain = self.callCabal2nix "cs-blockchain" ./specs/chain/hs {};
-              cs-core = self.callCabal2nix "cs-core" ./specs/core/hs {};
             };
           };
         };
