+++ specs/cabal.project
@@ -1,4 +1,3 @@
 packages: semantics/hs
           ledger/hs
           chain/hs
-          core/hs
