+++ specs/chain/hs/src/Chain/GenesisBlock.hs
@@ -10,7 +10,7 @@ import Data.ByteString.Lazy.Char8 (pack)
 import Data.Set (Set, fromList)
 import Ledger.Core (VKey(VKey), Owner(Owner), VKeyGenesis(VKeyGenesis))
 import Numeric.Natural
-import Chain.Types
+import Types
 
 -- | Verification keys located in the genesis block
 initVKeys :: Set VKeyGenesis
