--- specs/core/hs/cs-core.cabal
@@ -1,33 +1,0 @@
--- Initial cs-core.cabal generated by cabal init.  For further
--- documentation, see http://haskell.org/cabal/users-guide/
-
-name:                cs-core
-version:             0.1.0.0
-synopsis:            Core and common types and functionality for Cardano specifications
--- description:
-homepage:            https://github.com/input-output-hk/cardano-chain
-license:             MIT
-license-file:        LICENSE
-author:              IOHK Formal Methods Team
-maintainer:          formal.methods@iohk.io
--- copyright:
-category:            Testing
-build-type:          Simple
-extra-source-files:  ChangeLog.md
-cabal-version:       >=1.10
-
-library
-  exposed-modules:     Chain.GenesisBlock
-                     , Chain.Types
-                     , Ledger.Core
-                     , Ledger.Signatures
-  -- other-modules:
-  -- other-extensions:
-  build-depends:       base >=4.11 && <5
-                     , bytestring
-                     , containers
-                     , cryptonite
-                     , lens
-                     , memory
-  hs-source-dirs:      src
-  default-language:    Haskell2010
