--- specs/core/hs/default.nix
@@ -1,4 +1,0 @@
-let
-  pkgs = import ../../../pkgs.nix;
-in
-  pkgs.haskell.packages.ghc861.cs-core
