+++ specs/ledger/hs/cs-ledger.cabal
@@ -23,14 +23,15 @@ flag development
 
 library
   hs-source-dirs:      src
-  exposed-modules:     Ledger.Core.Generator
+  exposed-modules:     Ledger.Core
+                     , Ledger.Core.Generator
                      , Ledger.Delegation
+                     , Ledger.Signatures
                      , Ledger.UTxO
   build-depends:       base >=4.11 && <5
                      , bytestring
                      , containers
                      , cryptonite
-                     , cs-core
                      , hedgehog
                      , lens
                      , memory
@@ -55,7 +56,6 @@ test-suite ledger-delegation-test
   default-language:    Haskell2010
   build-depends: base
                , containers
-               , cs-core
                , lens
                , hedgehog
                , tasty
