+++ specs/ledger/hs/test/Ledger/Delegation/Examples.hs
@@ -11,7 +11,6 @@ import Numeric.Natural (Natural)
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit (testCase)
 
-import Chain.Types (DCert(DCert))
 import Ledger.Core
   ( Epoch(Epoch)
   , Owner(Owner)
@@ -24,6 +23,7 @@ import Ledger.Core
   )
 import Ledger.Delegation
   ( ADELEG
+  , DCert(DCert)
   , DSEnv(DSEnv)
   , DSState(DSState)
   , DState(DState)
